+++ source/SeqFlatFileImport.Tests/SeqFlatFileImport.Tests.csproj
@@ -3,8 +3,7 @@
   <PropertyGroup>
     <VersionPrefix>0.1.0</VersionPrefix>
     <TargetFramework>net6.0</TargetFramework>
-    <AssemblyName>Tests</AssemblyName>
-    <PackageId>Tests</PackageId>
+    <PackageId>SeqFlatFileImport.Tests</PackageId>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <GenerateAssemblyTitleAttribute>false</GenerateAssemblyTitleAttribute>
     <GenerateAssemblyDescriptionAttribute>false</GenerateAssemblyDescriptionAttribute>
