+++ source/SeqFlatFileImport.sln
@@ -6,7 +6,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SeqFlatFileImport", "SeqFla
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SeqFlatFileImport.Core", "SeqFlatFileImport.Core\SeqFlatFileImport.Core.csproj", "{86DEC7E9-4AC0-4166-B251-5E667C939E6E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{9BC81AEF-B4EC-4B39-A96B-71E7075CB62A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SeqFlatFileImport.Tests", "SeqFlatFileImport.Tests\SeqFlatFileImport.Tests.csproj", "{9BC81AEF-B4EC-4B39-A96B-71E7075CB62A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "_build", "..\build\_build.csproj", "{F5E1DE65-81E1-4B80-9EA2-42ED84877D7D}"
 EndProject
