--- source/Tests/LogFiles/Web-2017-03-28.log
@@ -1,32 +1,0 @@
-#Software: Octopus Deploy 0.0.0-local
-#Version: 1.0
-#Date: 2017-03-28 03:49:55
-#Fields: date time cs-method cs-uri-stem s-port cs-username c-ip sc-status time-taken
-2017-03-28	03:49:55	GET	/api	8065	-	::1	200	513
-2017-03-28	03:49:55	GET	/api/users/me	8065	-	::1	401	103
-2017-03-28	03:49:56	GET	/api/featuresconfiguration	8065	-	::1	401	139
-2017-03-28	03:49:56	GET	/api/serverstatus	8065	-	::1	401	204
-2017-03-28	03:49:56	GET	/api/dashboard	8065	-	::1	401	199
-2017-03-28	03:49:57	GET	/api/serverstatus	8065	-	::1	401	86
-2017-03-28	03:50:04	POST	/api/users/login	8065	-	::1	200	384
-2017-03-28	03:50:04	GET	/api/users/Users-1/permissions	8065	admin	::1	200	364
-2017-03-28	03:50:05	GET	/api/featuresconfiguration	8065	admin	::1	200	136
-2017-03-28	03:50:05	GET	/api/dashboard	8065	admin	::1	200	553
-2017-03-28	03:50:06	GET	/api/projects/Projects-1/logo	8065	admin	::1	304	424
-2017-03-28	03:50:07	GET	/api/serverstatus	8065	admin	::1	200	42
-2017-03-28	03:50:08	GET	/api/featuresconfiguration	8065	admin	::1	200	39
-2017-03-28	03:50:08	GET	/api/repository/configuration	8065	admin	::1	200	48
-2017-03-28	03:50:08	GET	/api/serverstatus	8065	admin	::1	200	39
-2017-03-28	03:50:08	GET	/api/serverstatus/nuget	8065	admin	::1	200	293
-2017-03-28	03:50:08	GET	/api/packages	8065	admin	::1	200	534
-2017-03-28	03:50:10	GET	/api/tasks	8065	admin	::1	200	1327
-2017-03-28	03:50:10	GET	/api/machines/all	8065	admin	::1	200	318
-2017-03-28	03:50:10	GET	/api/environments/all	8065	admin	::1	200	49
-2017-03-28	03:50:17	GET	/api/serverstatus	8065	admin	::1	200	25
-2017-03-28	13:50:31	GET	/api/serverstatus	8065	admin	::1	200	2915
-2017-03-28	13:50:42	GET	/api/serverstatus	8065	admin	::1	200	515
-2017-03-28	13:50:52	GET	/api/serverstatus	8065	admin	::1	200	21
-2017-03-28	13:50:57	GET	/api/dashboard	8065	admin	::1	200	33
-2017-03-28	13:51:02	GET	/api/serverstatus	8065	admin	::1	200	21
-2017-03-28	13:51:12	GET	/api/serverstatus	8065	admin	::1	200	21
-2017-03-28	03:51:22	GET	/api/serverstatus	8065	admin	::1	200	23
