--- source/Tests/LogFiles/u_ex150708.log
@@ -1,31 +1,0 @@
-#Software: Microsoft Internet Information Services 8.5
-#Version: 1.0
-#Date: 2015-07-08 23:01:55
-#Fields: date time s-ip cs-method cs-uri-stem cs-uri-query s-port cs-username c-ip cs(User-Agent) cs(Referer) sc-status sc-substatus sc-win32-status time-taken
-2015-07-08 23:01:55 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 485
-2015-07-08 23:01:57 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:01:58 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 12
-2015-07-08 23:01:59 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:00 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:01 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 3
-2015-07-08 23:02:02 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:02:03 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:04 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 0
-2015-07-08 23:02:05 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:06 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:07 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:08 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:09 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 3
-2015-07-08 23:02:10 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:11 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 0
-2015-07-08 23:02:12 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:02:13 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:02:14 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 1
-2015-07-08 23:02:15 ::1 GET /api - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:15:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:16:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 3
-2015-07-08 23:17:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:18:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:19:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
-2015-07-08 23:20:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 0
-2015-07-08 23:21:37 ::1 GET /api/dashboard - 80 - ::1 Mozilla/5.0+(Windows+NT+6.3;+WOW64)+AppleWebKit/537.36+(KHTML,+like+Gecko)+Chrome/43.0.2357.130+Safari/537.36 http://localhost:9005/octopus.dev.html?octopus.server=http%3A%2F%2Flocalhost%3A80%2F 404 0 2 2
